third_party/blink/web_tests/platform/generic/http/tests/websocket/no-subprotocol-expected.txt

WebSocket's protocol attribute should be an empty string if the server did not provide Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.protocol is ""
Connected
PASS ws.protocol is ""
Closed
PASS ws.protocol is ""
PASS closeEvent.wasClean is true
Exited onclose handler
PASS ws.protocol is ""
PASS successfullyParsed is true

TEST COMPLETE
